ironicclient/v1/conductor.py
# Licensed under the Apache License, Version 2.0 (the "License"); you may
# not use this file except in compliance with the License. You may obtain
# a copy of the License at
#
#      http://www.apache.org/licenses/LICENSE-2.0
#
# Unless required by applicable law or agreed to in writing, software
# distributed under the License is distributed on an "AS IS" BASIS, WITHOUT
# WARRANTIES OR CONDITIONS OF ANY KIND, either express or implied. See the
# License for the specific language governing permissions and limitations
# under the License.

from ironicclient.common import base
from ironicclient.common.i18n import _
from ironicclient.common import utils
from ironicclient import exc


class Conductor(base.Resource):
    def __repr__(self):
        return "<Conductor %s>" % self._info


class ConductorManager(base.Manager):
    resource_class = Conductor
    _resource_name = 'conductors'

    def list(self, marker=None, limit=None, sort_key=None, sort_dir=None,
             fields=None, detail=False):
        """Retrieve a list of conductors.

        :param marker: Optional, the hostname of a conductor, eg the last
                       conductor from a previous result set. Return
                       the next result set.
        :param limit: The maximum number of results to return per
                      request, if:

            1) limit > 0, the maximum number of ports to return.
            2) limit == 0, return the entire list of ports.
            3) limit param is NOT specified (None), the number of items
               returned respect the maximum imposed by the Ironic API
               (see Ironic's api.max_limit option).

        :param sort_key: Optional, field used for sorting.

        :param sort_dir: Optional, direction of sorting, either 'asc' (the
                         default) or 'desc'.

        :param fields: Optional, a list with a specified set of fields
                       of the resource to be returned. Can not be used
                       when 'detail' is set.

        :param detail: Optional, boolean whether to return detailed information
                       about conductors.

        :returns: A list of conductors.

        """
        if limit is not None:
            limit = int(limit)

        if detail and fields:
            raise exc.InvalidAttribute(_("Can't fetch a subset of fields "
                                         "with 'detail' set"))

        filters = utils.common_filters(marker, limit, sort_key, sort_dir,
                                       fields, detail)
        path = ''
        if filters:
            path += '?' + '&'.join(filters)

        if limit is None:
            return self._list(self._path(path), "conductors")
        else:
            return self._list_pagination(self._path(path), "conductors",
                                         limit=limit)

    def get(self, hostname, fields=None):
        return self._get(resource_id=hostname, fields=fields)
